Refatoring code to remove duplicated code
