+++ rundeckapp/grails-app/controllers/rundeck/controllers/EditOptsController.groovy
@@ -20,6 +20,7 @@ import org.apache.log4j.Logger
 import rundeck.Option
 import rundeck.ScheduledExecution
 import rundeck.services.FrameworkService
+import rundeck.utils.OptionsUtil
 
 import java.util.regex.Pattern
 import java.util.regex.PatternSyntaxException
@@ -561,7 +562,8 @@ class EditOptsController {
             if(opt.realValuesUrl){
                 try{
                     def realUrl = opt.realValuesUrl.toExternalForm()
-                    def urlExpanded = expandUrl(opt, realUrl, params.project, [:], realUrl.matches(/(?i)^https?:.*$/))
+                    ScheduledExecution se = opt.scheduledExecution
+                    def urlExpanded = OptionsUtil.expandUrl(opt, realUrl, se, [:], realUrl.matches(/(?i)^https?:.*$/))
                     def remoteResult=ScheduledExecutionController.getRemoteJSON(urlExpanded, 10, 0, 5)
                     if(remoteResult){
                         def remoteJson = remoteResult.json
@@ -579,82 +581,7 @@ class EditOptsController {
         return result
     }
 
-    /**
-     * Expand the URL string's embedded property references of the form
-     * ${job.PROPERTY} and ${option.PROPERTY}.  available properties are
-     * limited
-     */
-    protected static String expandUrl(Option opt, String url, String project,selectedoptsmap=[:],boolean isHttp=true) {
-        def invalid = []
-        def frameworkService = grails.util.Holders.applicationContext.getBean('frameworkService') as FrameworkService
-        def se = grails.util.Holders.applicationContext.httpSession
-        def rundeckProps=[
-                'nodename':frameworkService.getFrameworkNodeName(),
-                'serverUUID':frameworkService.serverUUID?:''
-        ]
-        if(!isHttp) {
-            rundeckProps.basedir= frameworkService.getRundeckBase()
-        }
-        def extraJobProps=[
-                'user.name': (se?.user?: "anonymous"),
-        ]
-        extraJobProps.putAll rundeckProps.collectEntries {['rundeck.'+it.key,it.value]}
-        Map globals=frameworkService.getProjectGlobals(project)
-
-        def replacement= { Object[] group ->
-            if (group[2] == 'job' && jobprops[group[3]] && scheduledExecution.properties.containsKey(jobprops[group[3]])) {
-                scheduledExecution.properties.get(jobprops[group[3]]).toString()
-            } else if (group[2] == 'job' && null != extraJobProps[group[3]]) {
-                def value = extraJobProps[group[3]]
-                value.toString()
-            }else if (group[2] == 'globals' && null != globals[group[3]]) {
-                def value = globals[group[3]]
-                value.toString()
-            }else if (group[2] == 'rundeck' && null != rundeckProps[group[3]]) {
-                def value = rundeckProps[group[3]]
-                value.toString()
-            } else if (group[2] == 'option' && optprops[group[3]] && opt.properties.containsKey(optprops[group[3]])) {
-                opt.properties.get(optprops[group[3]]).toString()
-            } else if (group[2] == 'option' && group[4] == '.value') {
-                def optname = group[3].substring(0, group[3].length() - '.value'.length())
-                def value = selectedoptsmap && selectedoptsmap instanceof Map ? selectedoptsmap[optname] : null
-                //find option with name
-                def Option expopt = scheduledExecution.options.find { it.name == optname }
-                if (value && expopt?.multivalued && (value instanceof Collection || value instanceof String[])) {
-                    value = value.join(expopt.delimiter)
-                }
-                (value ?: '')
-            } else {
-                null
-            }
-        }
-        //replace variables in the URL, using appropriate encoding before/after the URL parameter '?' separator
-        def arr=url.split(/\?/,2)
-        def codecs=['URIComponent','URL']
-        def result=[]
-        arr.eachWithIndex { String entry, int i ->
-            result<<entry.replaceAll(/(\$\{(job|option|rundeck|globals)\.([^}]+?(\.value)?)\})/) { Object[] group ->
-                def val = replacement(group)
-                if (null != val) {
-                    if(!isHttp){
-                        return val
-                    }
-                    val."encodeAs${codecs[i]}"()
-                } else {
-                    invalid << group[0]
-                    group[0]
-                }
-            }
-        }
-        String srcUrl = result.join('?')
-        if (invalid) {
-//            LogFactory.getLog(this).error("invalid expansion: " + invalid);
-            logger.error("invalid expansion: " + invalid)
-        }
-        return srcUrl
-    }
-
-    /**
+        /**
      * Use input parameters to configure an Option object.
      * Special properties "valuesType" and "enforcedType" configure
      * the option when mutually exclusive properties are in the input parameters.
